requirements.txt
docopt
pexpect
-e git+https://github.com/ftao/python-ifcfg.git#egg=ifcfg
web.py
pyzmq
